fix choppy audio output
I was using pyaudio and it was working well for some time, but since running it
on a newer system with python3 it got choppy. Lowering the frames_per_buffer to
1 improves it slightly, but doesn't eliminate the problem. Fortunately, at the
same time, using /dev/dsp via padsp stopped having an irritating amount of
latency, so that is what I'm using now.

I was thinking of removing the pyaudio and pacat codepaths completely, but
pyaudio might make it work on other OSes, and the pacat command could be
replaced with something else too, so I'm leaving both here for now. But if
you're using a modern PulseAudio system, I recommend padsp.